FIX: load models multiple times due to multiple inheritance of PredictFastaSpecLib

+++ peptdeep/model/ms2.py
@@ -722,8 +722,8 @@ def spearman_correlation(x: torch.Tensor, y: torch.Tensor, device):
         Shape (Batch, n)
 
     """
-    x_rank = _get_ranks(x, device)
-    y_rank = _get_ranks(y, device)
+    x_rank = _get_ranks(x, device).to(torch.float32)
+    y_rank = _get_ranks(y, device).to(torch.float32)
     
     n = x.size(1)
     upper = 6 * torch.sum((x_rank - y_rank).pow(2), dim=1)
